Migrated commands to towny admin, and improved error handling.
