+++ src/com/palmergames/bukkit/towny/command/PlotCommand.java
@@ -508,94 +508,6 @@ public class PlotCommand extends BaseCommand implements CommandExecutor {
 						player.sendMessage(TownySettings.getLangString("msg_err_empty_area_selection"));
 					}
 
-				} else if (split[0].equalsIgnoreCase("meta")) {
-
-					TownBlock townBlock = new WorldCoord(world, Coord.parseCoord(player)).getTownBlock();
-					plotTestOwner(resident, townBlock);
-					
-					if (split.length == 1) {
-						if (townBlock.hasMeta()) {
-							player.sendMessage(ChatTools.formatTitle("Custom Meta Data"));
-							for (CustomDataField field : townBlock.getMetadata()) {
-								player.sendMessage(field.getKey() + " = " + field.getValue());
-							}
-						} else {
-							TownyMessaging.sendErrorMsg(player, "This plot doesn't have any associated metadata");
-						}
-						
-						return true;
-					}
-					
-					if (split.length < 3) {
-						player.sendMessage(ChatTools.formatCommand("", "meta", "set", "The key of a registered data field"));
-						return false;
-					}
-					
-					if (split.length == 4) {
-						String operation = split[1];
-						String mdKey = split[2];
-						String val = split[3];
-
-						plotTestOwner(resident, townBlock);
-						
-						if (!townyUniverse.getRegisteredMetadataMap().containsKey(mdKey)){
-							TownyMessaging.sendErrorMsg(player, "The metadata for " + "\"" + mdKey + "\"" + " is not registered!");
-							return false;
-						} else if (split[1].equalsIgnoreCase("set")) {
-							CustomDataField md = townyUniverse.getRegisteredMetadataMap().get(mdKey);
-							
-							for (CustomDataField cdf: townBlock.getMetadata()) {
-								if (cdf.equals(md)) {
-									
-									// Change state
-									try {
-										cdf.isValidType(val);
-									} catch (InvalidMetadataTypeException e) {
-										TownyMessaging.sendErrorMsg(player, e.getMessage());
-										return false;
-									}
-									
-									cdf.setValue(val);
-									
-									// Let user know that it was successful.
-									TownyMessaging.sendMessage(player, ChatColor.GREEN + "Key " + mdKey + " was successfully updated to " + cdf.getValue());
-
-									// Save changes.
-									townyUniverse.getDataSource().saveTownBlock(townBlock);
-								}
-							}
-							
-							return true;
-							
-						} 
-					} else if (split[1].equalsIgnoreCase("add")) {
-						String mdKey = split[2];
-
-						if (!townyUniverse.getRegisteredMetadataMap().containsKey(mdKey)) {
-							TownyMessaging.sendErrorMsg(player, "The metadata for " + "\"" + mdKey + "\"" + " is not registered!");
-							return false;
-						}
-						
-						CustomDataField md = townyUniverse.getRegisteredMetadataMap().get(mdKey);
-						if (townBlock.hasMeta()) {
-							for (CustomDataField cdf: townBlock.getMetadata()) {
-								if (cdf.equals(md)) {
-									TownyMessaging.sendErrorMsg(player, ChatColor.GREEN + "Key " + mdKey + " already exists!");
-									return false;
-								}
-							}
-						}
-						
-
-						player.sendMessage(ChatColor.GREEN + "Custom data was successfully added to townblock!");
-
-						townBlock.addMetaData(md);
-
-						// Save.
-						townyUniverse.getDataSource().saveTownBlock(townBlock);
-					}
-					
-						
 				} else
 					throw new TownyException(String.format(TownySettings.getLangString("msg_err_invalid_property"), split[0]));
